thumbor: convert generate-thumbor-age-metrics to timer
Because we can't just execute the command using a pipe
we also need to create a wrapper script that runs the previous
script and pipes it into nc.

And since that uses the variable statsd host and port it needs to
be an erb template that uses the existing variables.

Though we should not create .sh files from templates so it's also
separated into a config file and the actual script that includes
the config.

Bug: T273673
Change-Id: I91b5920ce456384fb674c9098d8d9715fa5ec2e1