durum: switch to client-side UUID generation
In the current version of durum, we are generating UUIDs on the durum
server by generating a UUID in Python and serving it through Flask. As
we are just using the Flask and uWSGI layers for generating a UUID, in
order to simplify the durum setup, reduce code complexity, and to help
us scale better, we should generate UUIDs using the client's browser
instead.

Given the requirements of the durum check service -- a simple DNS lookup
for *.check.wikimedia-dns.org -- there are no security concerns with
doing something like this from client instead of the server. (The TTL
for *.check.wikimedia-dns.org is set to five seconds.)

To generate the UUIDs on the client side, we do not roll our own UUID
generator and instead use uuid from https://github.com/uuidjs/uuid. As
the uuidjs project does not publish an unminified version of uuidv4.js
and we don't want to ship that (for transparency of the JS we run on a
client's browser), we ship our own unminified version of uuidv4.js and
also document how to generate such a file, for reproducibility.

Bug: T289536
Change-Id: If59a37db0a08b60257ecfdb8250c6c61383c71a8
Hosts: O:durum