gitlab::backup remove deprication warning and deletion of config backup
In gitlab-ce 13.12.9 a deprication warning was introduced which we
had to grep out in I61b721fdf7ac061bb5dccb36e4189280d456bacc. With
GitLab 14 we don't need to filter out the warning. Config
backups are cleaned up similar to data backups in GitLab 14 as well.
So I removed the delete command.

Bug: T288324
Bug: T274463
Change-Id: I3cc43c16f48cf28e9ab4d6691a957770e72f2646