kubernetes: add revscoring-editquality in the services configs
The ML team would like to deploy the first helmfile-based service
definition on the deployment nodes.
There is already a draft related to how the helmfile config will
look like in https://gerrit.wikimedia.org/r/c/operations/deployment-charts/+/719128

Most of the work on which we'll focus during the next months
is related to add revscoring models, namely what we have on ORES.

Part of this change is to split tokens and secrets between "main"
clusters and "ml-serve" clusters, so that there is a clean separation
between the two.

For secrets, the idea is to create new directories on deploy1002 named:
- /etc/helmfile-defaults/private/ml-serve_services
- /etc/helmfile-defaults/private/main_services
- /etc/helmfile-defaults/private/admin_ng_services
The current /etc/helmfile/private's content is not absented, so any
deployment that may happen between merging this change and changing
helmfile configs in deployment-charts will not be disrupted. We will
eventually clean up the leftovers manually.

The admin token will be copied over from main to ml-serve for the moment
(so all clusters will keep sharing it) and a subsequent patch will change
it for ML clusters (configuring kube-api as well).

Before merging this change some changes to the puppet private repo
are of course needed.

Bug: T286791
Change-Id: I6cbdab0f868f62adb5fdeda2a632befca6db458f