query_service: Support proxying to microsite from backend
For WCQS trafficserver will be sending all requests to the nginx
fronting blazegraph (to support OAuth flow). Configure nginx to proxy
requests that don't resolve into the blazegraph api to a configured
url. For consistency purposes this applies equally to WDQS and WCQS,
but on WDQS this should effectively be a noop. trafficserver routes UI
requests for WDQS directly to the microsite. The one minor change to
WDQS is that 503 errors will now return the maintenance page from the
gui in all circumstances, not just when the ui is installed locally.

Hosts: wdqs1010.eqiad.wmnet,wcqs1001.eqiad.wmnet
Bug: T280247
Change-Id: I8a74d3443bde8baddf762129cebff98124f7ef02