cloud lvm: finish up volume group creation for ephemeral disk
This works in local testing. New systems with an ephemeral disk reliably
have a label on that disk of "ephemeral0" so far. This should allow
the usage of any ephemeral disk that is labeled like that to become a
PV and then a VG named "vd" which is all that is needed to make the rest
of the old labs_lvm module go.

Bug: T277078
Change-Id: I72177d778a738cb49840bd069f1e2571d6a98642