Grant access to anlytics-privatedata-users for Michael Raish
This change adds user 'mikeraish' to the anlytics-privatedata-users
group.  Additionally it sets user 'mraish' to be absent in data.yaml,
as it is a duplicate account no longer needed.

User 'mraish' was set up in April 2020 when Michael was contracting
with the foundation.  More recently, after going full time, a new
account, username 'mikeraish' was set up.  That is the account
Michael is currently using, so I've removed the older one from
LDAP and the reference to it in data.yaml.

Bug: T290766
Change-Id: I6f4a2f449fd1261f1079706e2b8d81c94cb8f3a6