profile::analytics::httpd: disable CGI
The old Wikistats v1 (previous version of stats.wikimedia.org)
is kept alive to allow links to work properly. In my opinion
we shouldn't allow the execution of old perl scripts that nobody
has touched in ages, and we shouldn't rely on mod_cgi|mod_cgid.

The profile is used by an-web1001, serving
{stats,analytics}.wikimedia.org, both sites not in need of CGI.

Bug: T285355
Change-Id: I354fd87d383304d4ce99bb3cc26e8a7b48aeb4f2