statistics: decom old httpd Directory for stats.wikimedia.org
The old /reportcard directories are going to be decommissioned,
they contain data from 2012 that is not used by anybody.
This is the first step to remove aliases and deny access to those
files. We'll then be able to remove/backup/etc.. files and
eventually clean all up (if needed).

This change also clean up the various httpd password files
(a follow up needs to happen on the private repo too).

Bug: T285355
Change-Id: Iecda0681f21d39bad566da0ed2dfeb9546863103