service::catalog: remove ProxyFetch checks from services on k8s
On k8s, we don't connect directly to the service, but rather to a
NodePort running on each node, which will in turn actually load-balance
traffic across pods and health checks (in the form of livenessProbe and
readinessProbe). The actual healthcheck from pybal is thus most likely
not going to be served by the current node and thus loses any meaning.

Also: fix some single quotes vs double quotes inconsistencies.

Change-Id: I3e72ed03925d0d2f771e4a261d5099bdfab1715e