apt::package_from_component: use apt-get update exec from init class
We already define a `apt-get update` exec in the apt class which is
included by this component.  As such it is safe to reuse that exec
definition for any dependency mapping. This also fixes a bug which
prevents users passing packages as hashs

Hosts: cumin:R:apt::package_from_component
Bug: T291370
Change-Id: I5120152ee3460dea2be6c91dd10563d10d0831b2