cloudnfs: set up a PoC Openstack instance-based nfs server
Instance-based nfs servers have always just been to test the hardware
systems until now. This copies much of the labstore project while dropping
things that don't work inside the cloud realm (mostly). It also assumes
we are not using DRBD for this machine.

As written, this will require special changes on the client-side puppet
code to actually mount anything such a server shares out. It also assumes
firewall will be handled with a security group that shares port 2049 to
cloud instances only.

I've pulled in the archive project vol and logcleanup script in case we
want to play with them? They appear to be abandoned work from the past.

Bug: T291406
Change-Id: I5e4400a033f5b3ce1c1bc97915e964ce0db2bad9