hieradata - O:puppetmaster::puppetdb: dont filter facts used for compilation
We need to ensure that facts used for catalog compilation are not
filtered from puppetdb.  The production puppetmaster doesn't need
theses values in the database for it to compile catalogs.  however when
we export facts to the PCC compilers we export the facts from puppetdb
as such any facts used in the production puppet repo should not be
filtered
 * numa is used by all hosts in realm.pp
 * cpu_details is used by the java::package class

Change-Id: I2c6318b1936e40897c2ed2a92511cf402f83ee3b