durum: add and set CSP headers for check.wikimedia-dns.org
Given the sensitivity of the durum check service, add and set
Content-Security-Policy headers for the landing page. We start by
setting the following policies: script-src, style-src, connect-src, and
frame-src, frame-ancestors. (default-src should cover script-src and
style-src but we want to be explicit.)

To properly use CSP and to avoid the use of unsafe-inline, we refactor
the code to move away from the inline stylesheet and script; these are
now referenced from an external file instead.

Why not use a CSP hash source to avoid setting unsafe-inline? Our code
is not minified for readability and transparency reasons, and generating
a checksum for code that includes whitespaces and may be updated later
is probably not a good strategy. (We would need to keep the hash
updated.)

Why not use a CSP nonce source? We are using nginx-light for durum,
which does not ship with the ngx_http_sub_module to allow us to perform
substitutions within the HTML file, once we generate a nonce. Then there
is the question of how to generate a nonce for this and that is best
avoided to keep the code simple.

So instead of using a hash, nonce, or unsafe-inline, we move the
stylesheet and the scripts to an external file to comply with CSP.

Since we are adding security headers, we also use this commit to set the
X-Content-Type-Options header.

Bug: T289536
Change-Id: I7e0ed378d23787508cf05ed2f9574a00fb897691
Hosts: O:durum