d-i: Stop using the udebs from sid for the bullseye config
During a netinst installation, the early d-i image is served via PXE/HTTP.
The image then pulls most of the installer components via minified deb
packages (so-called udebs). This includes most of the kernel modules.
As such, the version of the base kernel needs to match the version of the
udebs pulled during installation.

When we were preparing bullseye for WMF, Bullseye was still in flux
and there was often a mismatch between the last relesed d-i netinst image
and a kernel which was already updated in testing. To make this more
usable there's a d-i option to instead grab the udebs from unstable (to
avoid constant catchup) and which ensured that installations weren't
frequently interrupted.

However, when Bullseye became stable I forgot to revert this, which didn't
cause any immediate issues as testing stayed on 5.10 for a while, but with the
recent upgrade to 5.14 it now fails. So fix it for good by removing that
setting.

Change-Id: I166e003fca1b421561782907584be54e1dade646