profile::kuberentes_deployment_server: re-think user management
With the introduction of helm 3, we're starting to define more than one
user per namespace with possibly different user permissions.

Thus we restructure the way usernames are defined - we allow multiple
usernames per service, and inject tokens into the users data.

Credential files are then tied to the user, not the service itself.

Also to reduce repetition the hiera definition is centralized as we had
two almost exact copies in two different hiera files, and start using
defaults for users we can override but reduce the amount of identical
properties we define everywhere.

Finally, move the definitions of the kubernetes clusters to a hiera data
structure which encomprises them all.

Change-Id: I41b4c9522669b729c35207faa82a0fc984b81071