ldap::sssd: Don't specify services on bullseye
It turns out that you can either specify the services in the sssd
config or start them using sockets (systemd), and we want to use
sockets so we can mount them inside containers and such.

This is also a behavior change (it seems) as before it would not fail
to start the socket services even if they were specified in the sssd
config.

Hosts: re:.*(\.cloud)$
Bug: T291585
Change-Id: I5edfde40f4ac1519a5aa8a128fe8e4f69c5424e0
Signed-off-by: David Caro <dcaro@wikimedia.org>