wmflib::dir::mkdir_p: fix issue cause by trailing slash
Currently when users pass a trailing slash to this function we get a
failure created by a double resource definitions.

This happens because when we pass '/path/to/dir/' to the split function
we get back the following array
        ['/path, '/path/to, '/path/to/dir]
notice the last element has the slash removed, this when we try to
preform split($dirs) - $dirs nothing is removed as there is no match

This CR adds a new normalise function which will ensure all directories
in an array will have the trailing slash removed.  I have also added
additional spec tests to catch this error

Change-Id: Ib98462bd2fc2a0d152e813cfc2fa53d6a226fbfb