query_service: Split oauth secret from settings
All of the oauth settings had to go into the private repo, although
the only secret value was the consumer key. Split the consumer key out
of the hash at the hiera level and combine it into the expected
structure in the profile. Dropping the type specificity to Hash at
the profile level should be acceptable as it will be checked as an
OAuthSettings type when passed on.

As a follow-up to this, we should circle back to the /srv/private repo
to remove the non-consumer-key (non-secret) values.
We can do this after merging this patch and labs/private.

Additionally we need this to run PCC:
https://gerrit.wikimedia.org/r/c/labs/private/+/724831

And then this followup patch to clean up:
https://gerrit.wikimedia.org/r/c/labs/private/+/724832

Bug: T280006
Hosts: O:wcqs::public
Change-Id: I218e6cec3ede168f1486a1c284bf072e20b65666