geoip: create transitional class geoip::data::maxmind::ipinfo
To download Maxmind databases for the IP Info extension using
a different Maxmind license.

Use this transitional class to be able to download new databases in parallel
while avoiding duplicate definitions in puppet.

This merges in I3df3a7803bd15e0.

This class will be merged into the regular maxmind geoip class once we
fully migrated to the new license.

It is a copy of the existing class but it also uses a timer
insted of a cron, a change we wanted to do anyways, and of
course defaults to different log and data dir pathes.

It also includes Id5937150ef1662a now but just for the new files.

Bug: T273673
Bug: T288844
Change-Id: I4f6608f2789b87d96ca7c01146b053bd53911434