network: Split kubepods networks per cluster
Why:
The previous way of defining kubernetes pod and network entries wasn't
very multiple cluster friendly. It has served us up to now, but it's
time we move on.

What:
* Clearly define the "services/main" cluster svc/pods networks
* Clearly define the "staging" cluster svc/pods networks

Maintain compatibility with the old scheme for a while by keeping the
old vars and macros around but derive them from the new ones. Those
should be removed as followup patches

Amend urldownloader to use the new form instead of the old one

Change-Id: Id7a43a025c5e9c666354750d67c01bbc6f1f0f0c