aprepo: Import kubeadm to stretch too
I tried to be correct in I691eb68623627ab8543239fa2a86de682ed722f9.
Being correct is still a good idea, so import kubeadm to Stretch repos
for now. Note that this skips "thirdparty/kubeadm-k8s-docker.com" update
on purpose, as those are versioned and specific to each Debian release,
and we don't need them on Stretch.

Bug: T292131
Change-Id: Ibf4c4e8471461eec522b7bbe99ddad4ff28b1610