toolforge harbor: add external postgres db
This should allow for a replicated or unreplicated postgresql database
for harbor. It's not hard to run harbor from docker-compose, but if you
don't have shared state, you cannot recover very quickly or have any
HA. This will likely also require a redis cluster to have HA. However
if we just want recovery, then keeping the tarball in the cinder volume
for the installer and using the redis container is potentially sufficient.

Bug: T267616
Change-Id: I02f80695833b55fbd216e7a013214aae816c8c69