zuul: raise zuul queue alarm
Raise threshold for the Zuul gearman waiting queue. It has been quite
spammy for a while and those are usually transient (people sending
patches causing a burst).

Later on that can be moved to use the Grafana alert at
https://grafana-rw.wikimedia.org/d/000000322/zuul-gearman?tab=alert&editPanel=10&orgId=1
and use the Alert manager to issue the notifications which is T292284

Bug: T292284
Change-Id: Id867b7a9fd5376ea9d5d75d179903cd641440d06