alertmanager: add release engineering team
Add notifications boilerplate for our public IRC channel and the team
mailling list.
The sole rule is that anything tagged `team:releng` is emitting to IRC
channel. We can later expand to have finer rule.

I have made the Grafana panel for Zuul Gearman functions to tag its
alert with `team:releng`, so when it fires that should theorically emit
a notification to IRC.

Link to the dashboard:
https://grafana-rw.wikimedia.org/d/000000322/zuul-gearman?tab=alert&editPanel=10&orgId=1

Bug: T292284
Change-Id: I7325603d213efeab96c474ec396145d522cbb79e