AuthDNS DoTLS - use alternate LE chain
This is a server-facing service which faces the recdns servers of
the world, most-notably Google Public DNS.  It's a pretty safe bet
they're updated for the ISRG Root X1 as we're also requiring
TLSv1.3.  The excess chain link to the expired DST Root could
cause problems and has no potential benefits here, and we have
some minor indications of a possible ongoing related problem which
there's a small chance this change could fix!

Change-Id: I7939cb0e23ba50fc98627fd093cf98753df11736