wikistats: add email capability to import jobs
This should send email about the state of import jobs,
which are currently miraheze and neoseeker.

It will not affect the regular updating of wiki statistics
just the import of lists of wikis themselves.

First have to see if it would even detect a failure like
the one described on the ticket.

Will change email address from personal to a group of some
kind once verified.

Bug: T292369
Change-Id: I6852f299dafc626f2af29badc99887324ac7f56c