Add audio packages to support DSE hackathon on stat100x
One of the projects of the DSE hackathon is to use Tensorflow Magenta
for audio mashups. This change adds the necessary packages,
we'll decide later on if those are permanently needed
or if we want them only to support this hackathon (so as a one off).

More info https://github.com/magenta/magenta/blob/main/README.md#manual-install-wo-anaconda

Bug: T292306
Change-Id: I7b9aa1c37629dd5699b3fc0a741a4c3f4a9733e1