cache: exclude single backend experiment from pooled ATS backends
The list of *pooled* backends used by Varnish and rendered under
/etc/varnish/directors.frontend.vcl must not include the node on which
we are running the single-backend architecture experiment.

Considering that with 40a86a2 we already exclude the experimental node
from the list of backends *known* to Varnish, ie: the backends
definitions in /etc/varnish/wikimedia_text-frontend.vcl, we could in
theory simply depool the node we are experimenting with from the ats-be
keyspace. That does not seem like a particularly robust solution though,
in particular considering that we may be running the experiment for a
relatively long time.

To avoid introducing additional complexity to the already obscure Puppet
Ruby template rendering a Go template rendering a VCL file, use a string
called $confd_experiment_fqdn which is empty if the experiment is not
enabled on any host. This way, the Go template pipeline would read:

 if and (eq $data.pooled "yes") (ne $node "")

The latter part of which is true for all meaningful values of $node.

Bug: T288106
Hosts: cp3050.esams.wmnet
Change-Id: I8cdc9459478019d061918e614b69b1fa9c57cc1b