Revert "Revert "mediawiki/geoip: add option to also pull new MaxMind databases from master""
This reverts commit 94dbcd5c04a25f459407d4d804d99494c1803156.

Reason for revert: avoiding to ensure absent directory but also use recurse

In the original change, Ia22ad58534c99, we had a combination of
ensure => absent with recurse => true on all hosts not opting in.

This lead to puppet run failing on all hosts using mediawiki::common
except mwdebug1001. (https://phabricator.wikimedia.org/P17430)

With this revert-revert that issue should be fixed, we are simply
not creating the file resource if the parameter for it is not set to true.

Change-Id: I63a4e695464e45a233805dc306d24d7c6597743a