Increase the maximum renewable lifetime of a Kerberos ticket
We recently enabled an auto-renew service for Kerberos enabled users on
statistics servers. This reduced the frequency at which users had to
enter their Kerberos passwords from 2 days to 7 days.

This commit builds upon that change by increasing the maximum possible
lifetime of a ticket (as defined by the KDC) and the default renewable
lifetime of a ticket (specified by the requester).

The effect will be that users will not have to enter their password
any more often than every 14 days.

Bug: T268985
Change-Id: Ifb66d0f050ab0ecb54cc60998b24e481a4efa96b