bird: add IPv6 support to bird and anycast-healthchecker
This commit adds support for enabling IPv6 in bird as well as
anycast-healthchecker. Both these tools already support IPv6 but the
implementation for it was lacking in our puppetization as there was no
use case for it (as mentioned in the anycast documentation [0]).

With Wikidough running as any anycasted service but missing an IPv6
address, we finally have a requirement for adding IPv6 support to both
these tools and this commit is a step towards that. Once Wikidough has
its own isolated IPv6 address, we will use that as the first
implementation of this support.

[0] - https://wikitech.wikimedia.org/wiki/Anycast#Future_evolution

Bug: T292737
Hosts: O:wikidough
Change-Id: I2392fc6e77ef1bf2ea5cacfd922f475b5af2e2c3