toolforge harbor: puppetize the install/compose config file and such
This should help keep configuration in source control for the docker
setup used to launch this docker registry.

The docker-compose is set up with restart: always, so unless you stop
services with a `docker-compose down`, services should start at boot
and all is well. It's a different way of managing things, but it is
well-documented upstream and inline with several local dev environments.

Bug: T267616
Change-Id: Ib7a4b81eef500e3fa41ffd148c4eaf953d8fbf56