gitlab::ssh explicitly add git user with fixed id
The gitlab::ssh module does not add a git user explicitly. GitLab
adds the user while installing the package if it doesn't exist.
However GitLab creates the user with disabled login. Pulling over
SSH leads to the error "User git not allowed because account is
locked". This change explicitly creates the user and fixes the id
to make sure production GitLab and the replica use the same id.

Hosts: C:gitlab
Bug: T283076
Change-Id: I84199846673fab616c997c9c920520c5dda3d3b2