mediawiki::packages::fonts: replace fonts-liberation with fonts-liberation2
Liberation 2 produces better results at smaller sizes and is more
actively maintained than Liberation 1. Because liberation is the default
font for SVG text and most thumbnails are in the 200-500px range where
liberation1 does not have good hinting, it provides a bad user
experience. An example of rendering changes is provided in T253600.

Liberation 2 does not provide Liberation Sans Narrow, so Liberation 1
needs to stay installed. In the future, Liberation Sans Narrow may be
packaged seperately from the rest of Liberation 1.
Both Liberation 1 and 2 use the same font family name and other
properties, so users can not pick which versions to use. Liberation 2
has a higher `fontversion`, so it should be selected first by
fontconfig. Liberation 2 is licensed under the SIL OFL.

Bug: T253600
Bug: T247567
Change-Id: I8be5d1201264e0a471cacaed24802439616b3c57