hadoop: remove sudo usage in the check_hdfs_topology
The script doesn't need to use sudo since it is already handled
by the Nagios NRPE config. From local tests on the node this seems
the cause of an issue that we are seeing, namely the check returning
"OK" to alert1001 (incorrectly) but returning CRITICAL when called
locally (correctly).

My current suspicion is that the chain of sudo (outsite and inside
the script) lead to a failure in the hdfs dfsadmin command, that in
turns causes the return of "OK".

Bug: T292846
Change-Id: I660b4108eaee0aea1110c4e51466c04dd2b04044