Obsolete role::restbase::base
There's new Puppet code to designate a point of contact per role (which
will initially be used for OS updates, but most probably be extended
later).

Our Puppet styleguide expects every server to only have one role and
in order to avoid special hacks in the code which parses the point
of contact, let's remove the handful of remaining style guide violations.

Anything which gets included by role::restbase::base is already a profile,
so simply inline the two roles which use it.

Hosts: P:restbase
Change-Id: If36d568ffb03d18aad3b5b958c2d480cbae841ba