Dumps: Clarify licensing for Wikidata and update various links
Updated mission link as it moved from the governance wiki to the
Wordpress site. Clarify licensing of Wikidata, not all content in
Wikidata is public domain, only structured data.

Updated the link for the GFDL license as it linked to
Wikipedia:Copyrights which gives a general overview of Wikipedia
copyright which is not particularly useful or expected, just link to
the license text for GFDL like the other license links.

Updated the license for images paragraph as 4.0 is the default license
on enwiki and commons, and changed "many images are NOT released under
Creative Commons" to "many images are NOT released under this license"
as many images are released under different Creative Commons licenses
and versions, as well as different licenses.

Bug: T279436
Change-Id: Id4d0eae5ede51ace2ed7ab83ccc734d5415310b8