db-kill: Fix wmfmariadbpy package dependency
db-kill depends on python3-wmfmariadbpy for wmfmariadbpy.dbutil,
but that is not installed explicitly, wmfmariadbpy-admin is, which
depends on python3-wmfmariadbpy and installs it. As it is not
defined on puppet, it produces an error.

For now, change the dependency to wmfmariadbpy-admin, to avoid the
error maybe later, can be done in a more elegant way.

Change-Id: Ice23179f2be17cdd8827759032488581dad7ffb3