wikidough: update monitoring::service to check the host IP
We are currently monitoring the Wikidough service against the anycasted
IP. To ensure that the check is performed against the actual host and
not the anycasted IP, we should check the availability of the DoH and
DoT frontends against the host IPs instead.

Additionally we should add a check against the anycasted IP in
anycast_monitoring.pp but that's for a separate commit.

Hosts: O:wikidough
Change-Id: Ie891c0117deee5dbd70a82af99c5e04dc491807b