gitlab: deactivate new backup-restore unit for now
follow-up to Ib3de09daa31c

For some reason yet to be determined the backup-restore
script when running as a systemd service breaks
https://gitlab-replica.wikimedia.org because workhorse
gets killed and has to be manually started.

An issue that apparently was not happening when this script
was ran manually before it became a systemd timer/unit.

This is just to give time to properly debug without breaking
the website when the timer triggers the script.

Also will follow-up with a change to let us trigger this in Hiera.

Bug: T285867
Change-Id: I3e08d2c71c6ac5b78d0806d13ef6bd9df52235b4