gitlab: allow installing the restore script while NOT enabling the timer
follow-up to Ib3de09daa31c
follow-up to I3e08d2c71c6ac5b7

This is to (temporarily?) allow us to install the restore
script on the server without also activating the systemd timer
at the same time.

Current code did not allow doing this separately.

But we want to disable the timer/job/service without also completely
removing the script since that makes debugging hard.

And now that it's possible also disable it in codfw via Hiera where
it can be easily reactivated later.

Feel free to revert this later once confident you don't need it anymore
or let's do this in a more elegant way that lets us control both things
separately or keep it.

Bug: T285867
Change-Id: I6835a72ad576051e824a498b000cf631a835246b