cloudbackup: deploy cinder-backup service
The cinder-backup service is responsible for handling interactions with the
cinder backup API.

We plan to use cloudbackup hardware servers to store backups of selected cinder
volumes. For now, only cloudbackup2002.codfw.wmnet.

This patch:
 * enables the openstack repository on cloudbackup boxes
 * installs the cinder-backup debian package
 * deploys the cinder configuration file
 * ensures the cinder-backup service is running
 * ensures a LVM volume is defined and mounted (to host the actual data)
 * enables firewalling access to required ports on cloudcontrol from cloudbackup
 * enables firewalling access to required ports on the ceph farm from cloudbackup

We will need to figure out how to manage LVM volumes on cloudbackup servers for
this, probably on a separated patch.

Also, the cinder backup API policy filter should be reviewed to ensure strict
controls on who/how can the API be used. This is also for a later patch.

Bug: T292546
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I30627f6eda0f1cec8c1b8dc976d56a5fc357f35d