osm: convert common role to profile, avoid role inside role
Here is another case of a role included inside another role
similar to the case at I4413b18c9d5e3.

As described there our styleguide expects every node to just
have a single role and we want to map roles to contacts.

This should be a profile. Also move the inclusion of the
standard profile to the role.

Change-Id: I2d457d5d4121f7d8d8830435a31ea0bb4f559a8c