Remove alluxio resources from puppet
We have evaluated alluxio and found that it does not meet our needs.
Therefore this commit removes the manifests and resources from puppet.

The user and group account remains, but the user is removed from the
hadoop group, so it will lose its privilges. The user and group will be
removed in a subsequent commit.

Packages will be purged manually from the two hosts where they were
installed.

Bug: T266641
Change-Id: Ic144d19fcaf30cc60477f488dd1fbbcd66c363f8