Filter prom-exported NEL stats to <=10min old reports
When we export a counter of the number of NEL reports received in
the past minute, we should also filter on the age of reports.
Browsers may buffer reports for many reasons; only ones that were
generated in the recent history are likely to be actionable for us.

Ten minutes seems like a fine threshold for relevancy.

This will take some noise out of NEL-based paging -- including
preventing the amusing but unfortunate case of being paged because
you fixed an issue that blocked NELs from being sent by some users.

Bug: T257527
Change-Id: Ibe2bd1582b5291ea71c97b14937740a2541b5733