Add rate of high-signal NELs as a status page metric
When users can't reach our sites because of issues in intermediate
networks like transit providers, we commonly see a sharp increase in
the incoming rate of two flavors of NEL reports: tcp.timed_out and
tcp.address_unreachable.

We started paging SRE on these same flavors of report with fb4c410;
we de-noised them a bit by limiting to only recently-generated reports
with 338b6dc.  Now let's show them to status page viewers as well, so
that the next time we see an incident like https://w.wiki/4DxC it's
obvious to everything that something is amiss.

In the case of the 2021-10-08 incident, the statuspage graph would
look something like this: https://phabricator.wikimedia.org/F34690458

Bug: T285569
Bug: T257527
Change-Id: I391867c067477bafaa504988869df50babfd36ab