profile::prometheus::alerts: update kafka mirror maker settings
There are currently some warnings in icinga for mirror maker's
throughput main-codfw -> main-eqiad. Since codfw is not the active
DC, it makes sense that it doesn't produce 100 msg/s on avg, but something
way lower. On the contrary, eqiad produces a lot of msgs, and 25 msg/s
seems to be a low bar.

Change-Id: I06958bb68bb9aceb8b6c01e6f1400d8aeea94e45