systemd::timer::job: add support for splay
I have the use case to run a job on a daily basis but stretch its
execution over a period of time. It would be executed on several WMCS
instances which might end up running on the same underlying hardware and
could cause some IO stress.

That can be achieved with systemd::timer splay parameter, add it to
systemd::timer::job (default undef) so we can later define a job that
runs daily but at no fixed time.

Add a spec to ensure that when called with splay parameter, we pass it
to the underlying systemd::timer.

Change-Id: Ibedbeaf8d1e4424d6635f10eaa61108d50760773