contint: regularly prune docker material
When changing a Jenkins job to use a new material, we do not prune all
images. They tend to pill up until the Docker partition is filed at
which point we have an outage for that specific node.

We also have some containers staying idle despite `docker stop` being
executed for all containers that run a specific build. Possibly cause
the Jenkins publishers might not be exected when a build is abruptly
aborted.

On Sunday, prune all images and volumes.
On other days only prune dangling images.

The jobs are stretched over an hour (splay=3600) to avoid a potential
spike of IO.

Hosts: P:ci::docker
Bug: T292729
Change-Id: I0ef00ac8a1e37af0c370cd615b3d9b2c49999c94