p::s::timesyncd: use ensure in auto_restarts
This is a followup to:

I94e01b25900b8a2e42386d23b59fc88988ad5638

There's a $ensure param for this class, which the dnsbox profiles
use to disable timesyncd.  The new auto_restarts part needs to
honor it as well, or we end up with a degraded systemd state on
these hosts.

Change-Id: Ia896bc1a34d3dc348ab82ebb72f8f5ec1a446c34