gitlab: remove cas3 from external providers
Shortly after announcing that gitlab.wikimedia.org was open to all
users, I noticed that new signups were being marked as External.  This
should prevent that.

gitlab-ansible-test lacks this value, and doesn't exhibit the behavior.

Background:

https://docs.gitlab.com/ee/integration/omniauth.html#configure-omniauth-providers-as-external

As followup, we'll need to uncheck the "External" box for any users
who've signed up in the meanwhile.  I'll handle that.

Bug: T293696
Change-Id: I89edf6ae8af7932e3dcc7c0a000444e6bae99336