Add the data-engineering team to Alertmanager
This commit is the first step in onboarding the Data Engineering
(previously Analytics) team to Alertmanager. A team route is added,
along with receivers for severity levels, page, critical, warning, and
task.

Note that the Data Engineering team is still in the process of migrating
some items to the new name, so elements such as the IRC channel and the
VictorOps/Splunk routing key continue to use 'analytics' and will be
changed at a later date.

Bug: T293399
Change-Id: I8bbb57ba75ce9991be87c849fbc3877a1c50256f