mediawiki::appserver: fetch additional MaxMind databases on all appservers
This is enabling on all appservers what is already enabled on canary
appservers.

After that was not an issue now deploying to all of production.

What this will do is create a new directory /usr/share/GeoIPInfo/ and
then pull GeoIP2-Anonymous-IP.mmdb and GeoIP2-Enterprise.mmdb into it.

From the volatile directory on the puppetmaster,
not directly from MaxMind.

Existing path /usr/share/GeoIP should not be touched by this in any way.

You can compare any canary appserver, f.e. mwdebug1001 and see it
already applied there.

This should be the last step of this ticket, minus the question how
we will do this on k8s.

Bug: T288844
Change-Id: I58688a882549f0180d8705080d931f35532d0695