cumin: add alias for gitlab and add that to misc-releng
This adds a new alias for gitlab hosts to cumin.

So you can use A:gitlab with cumin hosts to run actions
on all gitlab hosts using the puppet role called gitlab.

Without having to update it or add hostnames ever in
the future.

Then add the new alias to the meta-alias misc-releng which
matches that and gerrit and others.

Change-Id: I3b89adcca11a3845f82fed3f499f3b513cd02a22