backups: Send db2098:3318 to dbprov2002
This is a follow up of
https://gerrit.wikimedia.org/r/c/operations/puppet/+/731404 which was a
copy of https://gerrit.wikimedia.org/r/c/operations/puppet/+/721288.

The problem was that the on the original patch, it was forgotten to
change the destination dbprov, so the snapshot was failing cause
xtrabackup was trying to prepare a 10.4 version with a 10.1 version (as
dbprov2001 is stretch).
dbprov2002 is buster and should be the one doing all that.

I have manually run the backup changing the .cnf file to point it to
dbprov2002 and it has gone fine.

Bug: T290868
Bug: T280979
Change-Id: I420613e2a1e30647c1ea02dd8fdbf13ccbba5cbf