cloud: nfs: refresh exclude pattern for nagios check_disk
Parameters are fed to nagios NRPE check_disk plugin using
profile::monitoring::nrpe_check_disk_options. In NFS servers the `-i /exp/.*`
argument expands to `-i /exp/. /exp/..` thus polluting the command line and
making the trailing `--exclude-type=tracefs` useless.

This results in the check_disk plugin trying to check /sys/kernel/debug/tracing
which is not available for the nagios user. The plugin therefore reports a
bogus problem with the disk.

I think we can just drop the glob, it should be OK to ignore the
whole /exp/ dir.

Bug: T294086
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I4a0d9a956871bb31b3c55bcb817707a6a4bb1907