Add a temporary firewall rule to support cassandra3 migration
We need to run sstableloader on an-presto1001.eqiad.wmnet and load the
snapshots to the aqs_next cluster. However sstableloader requires direct
access to port 7000 of each instance.

The drange parameter is not supplied here, because the access is
required to the secondary IPs of the aqs_next hosts.

As with the rsync rules, this temporary firewall rule will be removed
once the migration is complete.

Bug: T291472
Change-Id: Ie2cba8e4bfb3dd0b209c887297835ae3c73c554d