airflow - Allow access to webserver port
This will allow folks in the analytics vlan to
access the airflow webserver UI without having to have ssh access
to the box itself (they still need to ssh tunnel through e.g. a stat box).

In the future we'd like to enforce some better authentication with the airflow
UI, perhaps using LDAP, but for now in the experimental airflow setup we have
this is fine.

Change-Id: I241cef6a518c30548308b09883ec3ad2ee817f6a