gitlab: disable puppet in script and change timer
Disabling Puppet during the backup restore because some gitlab
services have to be stopped as the restore process is running
but if this clashes with a scheduled Puppet run, the services
will be restarted as a corrective measure causing unwanted
side effects. Also changing the timer since it ran to close
to the backup process so it might result in failed restore
if the backup and replication processes had not completed.

Bug: T285867
Change-Id: I149d37beb04594fd2207d1985bd6997165d4539e