puppetmaster: reorder private hiera lookup
The host hiera overrides should come before site ones.

Before this patch:
 0) other overrides, including public hiera
 1) private site
 2) private host

With this patch:
 0) private host
 1) other overrides, including public hiera
 2) private site

This mirrors the behavior to public hiera, in which host overrides
are the highest priority.

Moreover, move the private host overrides to be trully highest priority,
otherwise there is no way to override data set in the public repo.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I81f3b5d107498a82d410a0159c22a8eb9b94c6b7